+++ test/fs/fs.html
@@ -5,14 +5,13 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>fs test</title>
-    <script src="../../system/fs.js"></script>
-    <script src="../../system/fs-plus.js"></script>
     <script src="https://cdn.jsdelivr.net/gh/kirakiray/ofa.js/test/ok.js"></script>
   </head>
   <body>
-    <script src="./base-test.js"></script>
-    <script src="./remove-test.js"></script>
-    <script src="./proxy-test.js"></script>
-    <script src="./plus-test.js"></script>
+    <script src="./base-test.js" type="module"></script>
+    <script src="./remove-test.js" type="module"></script>
+    <script src="./proxy-test.js" type="module"></script>
+    <script src="./normal-test.js" type="module"></script>
+    <script src="./getall-test.js" type="module"></script>
   </body>
 </html>
